% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame.FrameFit}
\alias{as.data.frame.FrameFit}
\title{Extract summary stats from FrameFit object}
\usage{
\method{as.data.frame}{FrameFit}(object, ...)
}
\arguments{
\item{object}{an object of type FrameFit.}

\item{...}{Other arguments.}
}
\description{
Extract summary stats from FrameFit object
}
