% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_model.R
\name{autoplot.FrameModel}
\alias{autoplot.FrameModel}
\title{Plot \code{FrameModel} using ggplot2}
\usage{
\method{autoplot}{FrameModel}(object, x = NULL, ...)
}
\arguments{
\item{object}{an object of type FrameModel.}

\item{x}{data matrix of isotopic measurements (Default NULL).}

\item{...}{other arguments.}
}
\description{
Plot an object of type \code{FrameModel} using ggplot2. Dual isotope plots will be generated and if more than 2 measurement dimensions are present a list of plots will be returned.
}
