% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.FrameFit}
\alias{coef.FrameFit}
\title{Extract coefficient estimates from \code{FrameFit} object}
\usage{
\method{coef}{FrameFit}(object, point.est = "mean", ...)
}
\arguments{
\item{object}{an object of type \code{FrameFit}.}

\item{point.est}{string, type of point estimate to compute. One of "mean|median" (Default "mean").}

\item{...}{other arguments.}
}
\description{
Extracts point estimates based on posterior means or medians from object of type \code{FrameFit}.
}
