% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dgp.R
\name{fit_dgp}
\alias{fit_dgp}
\title{Fit Dirichlet-Gaussian process prior model}
\usage{
fit_dgp(
  model,
  x,
  t,
  sd = data.frame(),
  eta = 0,
  rho = 0.3,
  rho.r = 0.6,
  estim.rho = FALSE,
  iter = 2000,
  cores = 1,
  chains = 4
)
}
\arguments{
\item{model}{Object of type FrameModel.}

\item{x}{Data frame corresponding to isotopic measurements.}

\item{t}{Vector corresponding to time points.}

\item{sd}{Data frame corresponding to measurement errors (Default empty data.frame).}

\item{eta}{numeric or vector, measurement noise magnitude (Default 0).}

\item{rho}{Numeric, scaled correlation length for source contributions (Default 0.3).}

\item{rho.r}{Numeric, scaled correlation length for fractionation (Default 0.6).}

\item{estim.rho}{Logical, if TRUE then correlation length is fitted via hierarchical model (Default FALSE).}

\item{iter}{Numeric, number of sampler iterations (Default 2000).}

\item{cores}{Integer, number of cores to use (Default = 1).}

\item{chains}{Integer, number of chains to use (Default = 4).}
}
\description{
Fit Dirichlet-Gaussian process prior model
}
