% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_frame.R
\name{fit_frame}
\alias{fit_frame}
\title{Fit independent time step model}
\usage{
fit_frame(model, x, t = NULL, sd = data.frame(), eta = 0, iter = 2000)
}
\arguments{
\item{model}{Object of type FrameModel.}

\item{x}{Data frame corresponding to isotopic measurements.}

\item{t}{Vector of time points (Default NULL).}

\item{sd}{Data frame corresponding to measurement errors (Default empty data.frame).}

\item{eta}{numeric or vector, measurement noise magnitude (Default 0).}

\item{iter}{Numeric, number of sampler iterations (Default 2000).}
}
\description{
Fit independent time step model
}
