% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gp.R
\name{fit_gp}
\alias{fit_gp}
\title{Fit Gaussian process prior on measurements}
\usage{
fit_gp(
  model,
  x,
  t,
  sd = data.frame(),
  eta = 0,
  rho = 0.5,
  iter = 2000,
  cores = 1,
  chains = 4
)
}
\arguments{
\item{model}{an object of type \code{FrameModel}.}

\item{x}{data frame with \eqn{d} columns corresponding to isotopic measurements.}

\item{t}{vector corresponding to scaled time points with same length as data frame \code{x}.}

\item{sd}{data frame corresponding to measurement errors with same shape as \code{x} (Default empty data.frame).}

\item{eta}{numeric or vector of length \eqn{d}, measurement noise magnitude overall or per isotopic measurement (Default 0).}

\item{rho}{numeric, scaled correlation length for Gaussian process (Default 0.5).}

\item{iter}{integer, number of sampler iterations (Default 2000).}

\item{cores}{integer, number of cores to use (Default = 1).}

\item{chains}{integer, number of chains to use (Default = 4).}
}
\description{
Fit time series model using Gaussian process priors on measurements and otherwise fitting source contribution and fractionation weights independently.
}
\details{
See \link{frame_model} for description of model dimensions \eqn{d}, \eqn{K} and \eqn{L}.
}
\examples{
# Define sources and fit the model
model <- frame_model(n2o_sources, n2o_frac)
fit <- fit_gp(model, n2o_test[,-1], t = n2o_test$t, eta = 5, chains = 1)
coef(fit)

# Using only 2 sources and 2 isotopic measurements
model2 <- frame_model(n2o_sources[1:2,c(1:2,4:5)], n2o_frac[,c(1:2,4:5)])
fit2 <- fit_gp(model2, n2o_test2[,-1], t = n2o_test2$t, eta = 5, chains = 1)
coef(fit2)
}
