% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_map.R
\name{isotope_map}
\alias{isotope_map}
\title{Isotope Mapping computes the source contribution and fractionation weights as the solution to a linear system of equations.}
\usage{
isotope_map(model, x)
}
\arguments{
\item{model}{Object of type FrameModel}

\item{x}{Data frame of \code{d} columns corresponding to isotopic measurements}
}
\description{
Isotope Mapping computes the source contribution and fractionation weights as the solution to a linear system of equations.
}
