% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n2o_frac}
\alias{n2o_frac}
\title{Nitrous oxide fractionation effect for reduction}
\description{
Nitrous oxide fractionation effect for reduction
}
\references{
L. Yu, E. Harris, D. Lewicka-Szczebak, et al., “What can we learn from N2O isotope data? – analytics, processes and modelling,” Rapid Communications in Mass Spectrometry, vol. 34, no. 20, Aug. 2020. \url{https://doi.org/10.1002/rcm.8858}
}
