% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n2o_sources}
\alias{n2o_sources}
\title{Nitrous oxide source data taken from Yu et al.}
\description{
Nitrous oxide source data taken from Yu et al.
}
\references{
\url{https://doi.org/10.1002/rcm.8858}
}
