% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n2o_test}
\alias{n2o_test}
\title{Simulated data of nitrous oxide isotopic measurements \eqn{\delta^{15}}N, \eqn{\delta^{15}}N$\eqn{^{SP}}, \eqn{\delta^{18}}O generated assuming contributions from nitrification, bacterial denitrification and fungal denitrification in addition to reduction.}
\description{
Simulated data of nitrous oxide isotopic measurements \eqn{\delta^{15}}N, \eqn{\delta^{15}}N$\eqn{^{SP}}, \eqn{\delta^{18}}O generated assuming contributions from nitrification, bacterial denitrification and fungal denitrification in addition to reduction.
}
