% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n2o_test2}
\alias{n2o_test2}
\title{Simulated data of nitrous oxide isotopic measurements \eqn{\delta^{15}}N, \eqn{\delta^{15}}N$\eqn{^{SP}} assuming contributions from nitrification and bacterial denitrification in addition to fractionation.}
\description{
Simulated data of nitrous oxide isotopic measurements \eqn{\delta^{15}}N, \eqn{\delta^{15}}N$\eqn{^{SP}} assuming contributions from nitrification and bacterial denitrification in addition to fractionation.
}
