% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n2o_weights}
\alias{n2o_weights}
\title{Fixed weights \code{f1}, \code{f2}, \code{f3} and \code{r} used to simulate the data set \code{n2o_test}.}
\description{
Fixed weights \code{f1}, \code{f2}, \code{f3} and \code{r} used to simulate the data set \code{n2o_test}.
}
