% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_map.R
\name{param_map}
\alias{param_map}
\title{Computes measurement means from a vector or data frame of source contribution and fractionation parameters}
\usage{
param_map(model, f)
}
\arguments{
\item{model}{Object of type FrameModel}

\item{f}{Data frame of \code{K+L} columns corresponding to source contributions and fractionation}
}
\description{
Computes measurement means from a vector or data frame of source contribution and fractionation parameters
}
