% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame.FrameFit}
\alias{as.data.frame.FrameFit}
\title{Extract summary statistics of \code{FrameFit} object as data frame.}
\usage{
\method{as.data.frame}{FrameFit}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an object of type \code{FrameFit}.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and converting column names (to syntactic names: see \code{make.names}) is optional. Note that all of \emph{R}'s \emph{base} package \code{as.data.frame()} methods use \code{optional} only for column names treatment, basically with the meaning of \verb{data.frame(*, check.names = !optional)}. See also the \code{make.names} argument of the \code{matrix} method.}

\item{...}{other arguments.}
}
\value{
Data frame having a column \code{variable} for the variable description names, \code{index} if no time points were supplied or \code{t} with time points, \code{mean} for the posterior means, \code{se_mean} indicating estimation error based on effective sample size, \code{sd} for posterior standard deviation, a list of quantiles including the median 50\% quantile, \code{n_eff} with effective sample sizes according to the Stan documentation, and \code{Rhat} the Gelman-Rubin statistic according to the Stan documentation.
}
\description{
Collects summary statistics from the \code{FrameFit} object and arranges them in a data frame.
}
