% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.FrameFit}
\alias{autoplot.FrameFit}
\title{Plot FrameFit object using ggplot2}
\usage{
\method{autoplot}{FrameFit}(object, ...)
}
\arguments{
\item{object}{an object of type FrameFit.}

\item{...}{Other arguments.}
}
\description{
Plot FrameFit object using ggplot2
}
