% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_model.R
\name{autoplot.FrameModel}
\alias{autoplot.FrameModel}
\title{Plot FRAME model using ggplot2}
\usage{
\method{autoplot}{FrameModel}(model, x = NULL, ...)
}
\arguments{
\item{model}{an object of type FrameModel.}

\item{x}{data matrix of isotopic measurements (Default NULL).}
}
\description{
Plot FRAME model using ggplot2
}
