% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dgp.R
\name{fit_dgp}
\alias{fit_dgp}
\title{Fit Dirichlet-Gaussian process prior model}
\usage{
fit_dgp(model, x, t, sd = data.frame(), rho = 0.1, rho.r = 0.2, iter = 2000)
}
\arguments{
\item{model}{Object of type FrameModel.}

\item{x}{Data frame corresponding to isotopic measurements.}

\item{t}{Vector corresponding to time points.}

\item{sd}{Data frame corresponding to measurement errors (Default empty data.frame).}

\item{rho}{Numeric, scaled correlation length for source contributions (Default 0.1).}

\item{rho.r}{Numeric, scaled correlation length for fractionation (Default 0.2).}

\item{iter}{Numeric, number of sampler iterations (Default 2000).}
}
\description{
Fit Dirichlet-Gaussian process prior model
}
