% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dgp.R
\name{fit_dgp}
\alias{fit_dgp}
\title{Fit Dirichlet-Gaussian process prior model}
\usage{
fit_dgp(
  model,
  x,
  t,
  sd = data.frame(),
  eta = 0,
  rho = 0.3,
  rho.r = 0.6,
  estim.rho = FALSE,
  iter = 2000,
  cores = 1,
  chains = 4
)
}
\arguments{
\item{model}{an object of type \code{FrameModel}.}

\item{x}{data frame with \eqn{d} columns corresponding to isotopic measurements.}

\item{t}{vector corresponding to scaled time points with same length as data frame \code{x}.}

\item{sd}{data frame corresponding to measurement errors with same shape as \code{x} (Default empty data.frame).}

\item{eta}{numeric or vector of length \eqn{d}, measurement noise magnitude overall or per isotopic measurement (Default 0).}

\item{rho}{Numeric, scaled correlation length for source contributions (Default 0.3).}

\item{rho.r}{Numeric, scaled correlation length for fractionation (Default 0.6).}

\item{estim.rho}{Logical, if TRUE then correlation length is fitted via hierarchical model (Default FALSE).}

\item{iter}{integer, number of sampler iterations (Default 2000).}

\item{cores}{integer, number of cores to use (Default = 1).}

\item{chains}{integer, number of chains to use (Default = 4).}
}
\description{
Fit time series model using Dirichlet-Gaussian process priors on source contribution and fractionation weights. The correlation lengths can be set individually and optionally estimated using a hierarchical model.
}
\details{
See \link{frame_model} for description of model dimensions \eqn{d}, \eqn{K} and \eqn{L}.
}
