% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_frame.R
\name{fit_frame}
\alias{fit_frame}
\title{Fit independent time step model}
\usage{
fit_frame(
  model,
  x,
  t = NULL,
  sd = data.frame(),
  eta = 0,
  iter = 2000,
  cores = 1,
  chains = 4
)
}
\arguments{
\item{model}{an object of type \code{FrameModel}.}

\item{x}{data frame with \eqn{d} columns corresponding to isotopic measurements.}

\item{t}{vector corresponding to scaled time points with same length as data frame \code{x} (Default NULL).}

\item{sd}{data frame corresponding to measurement errors with same shape as \code{x} (Default empty data.frame).}

\item{eta}{numeric or vector of length \eqn{d}, measurement noise magnitude overall or per isotopic measurement (Default 0).}

\item{iter}{integer, number of sampler iterations (Default 2000).}

\item{cores}{integer, number of cores to use (Default = 1).}

\item{chains}{integer, number of chains to use (Default = 4).}
}
\description{
Fit a time series model using FRAME at each point in time independently.
}
\details{
See \link{frame_model} for description of model dimensions \eqn{d}, \eqn{K} and \eqn{L}.
}
\examples{
# Define sources and fit the model
model <- frame_model(n2o_sources, n2o_frac)
fit <- fit_frame(model, n2o_test[,-1], eta = 5)
coef(fit)

# Using only 2 sources and 2 isotopic measurements
model2 <- frame_model(n2o_sources[1:2,c(1:2,4:5)], n2o_frac[,c(1:2,4:5)])
fit2 <- fit_frame(model2, n2o_test2[,-1], eta = 5)
coef(fit2)
}
