% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_frame.R
\name{fit_frame}
\alias{fit_frame}
\title{Fit independent time step model}
\usage{
fit_frame(model, x, sd = data.frame(), iter = 2000)
}
\arguments{
\item{model}{Object of type FrameModel.}

\item{x}{Data frame corresponding to isotopic measurements.}

\item{sd}{Data frame corresponding to measurement errors.}

\item{iter}{Numeric, number of sampler iterations.}
}
\description{
Fit independent time step model
}
