% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit generalized linear model model with spline bases}
\usage{
fit_glm(model, x, t, sd = data.frame(), M = 8, M.r = 4, iter = 2000)
}
\arguments{
\item{model}{Object of type FrameModel.}

\item{x}{Data frame corresponding to isotopic measurements.}

\item{t}{Vector corresponding to time steps.}

\item{sd}{Data frame corresponding to measurement errors (Default empty data.frame).}

\item{M}{Integer, number of degrees of freedom for source contributions (Default 8).}

\item{M.r}{Integer, number of degrees of freedom for fractionation (Default 4).}

\item{iter}{Numeric, number of sampler iterations.}
}
\description{
Fit generalized linear model model with spline bases
}
