% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gp.R
\name{fit_gp}
\alias{fit_gp}
\title{Fit Gaussian process prior on measurements}
\usage{
fit_gp(model, x, t, sd = data.frame(), rho = 0.1, iter = 2000)
}
\arguments{
\item{model}{Object of type FrameModel.}

\item{x}{Data frame corresponding to isotopic measurements.}

\item{t}{Vector corresponding to time points.}

\item{sd}{Data frame corresponding to measurement errors (Default empty data.frame).}

\item{rho}{Numeric, scaled correlation length for Gaussian process (Default 0.1).}

\item{iter}{Numeric, number of sampler iterations (Default 2000).}
}
\description{
Fit Gaussian process prior on measurements
}
