% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stationary.R
\name{fit_stationary}
\alias{fit_stationary}
\title{Fit stationary FRAME model}
\usage{
fit_stationary(model, x, iter = 2000)
}
\arguments{
\item{model}{Object of type FrameModel}

\item{x}{Data frame corresponding to isotopic measurements}

\item{iter}{Number of sampler iterations (Default 2000)}
}
\description{
Fit stationary FRAME model
}
