% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stationary.R
\name{fit_stationary}
\alias{fit_stationary}
\title{Fit stationary FRAME model}
\usage{
fit_stationary(model, x, eta, iter = 2000)
}
\arguments{
\item{model}{an object of type FrameModel.}

\item{x}{a data frame corresponding to isotopic measurements.}

\item{eta}{numeric or vector, measurement noise magnitude.}

\item{iter}{integer, number of sampler iterations (Default 2000)}
}
\description{
Fit stationary FRAME model
}
