% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_model.R
\name{print.FrameModel}
\alias{print.FrameModel}
\title{Print FRAME model}
\usage{
\method{print}{FrameModel}(x, ...)
}
\arguments{
\item{x}{an object of type \code{FrameModel}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print an object of type FrameModel.
}
