% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_map.R
\name{sample_measurements}
\alias{sample_measurements}
\title{Sample isotopic measurements from true parameter values}
\usage{
sample_measurements(model, f, eta)
}
\arguments{
\item{model}{an object of type \code{FrameModel}.}

\item{f}{vector with \eqn{K+L} components or data frame with \eqn{K+L} columns containing true parameter values.}

\item{eta}{numeric or vector of length \eqn{d}, measurement noise overall or per isotopic measurement.}
}
\description{
Uses distributions of source isotopic signature and fractionation factor in object of type \code{FrameModel} to sample parameters and then simulates measurements from them using noise defined in \code{eta}.
}
\details{
See \link{frame_model} for description of model dimensions \eqn{d}, \eqn{K} and \eqn{L}.
}
