% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_map.R
\name{sample_measurements}
\alias{sample_measurements}
\title{Sample isotopic measurements from true parameter values}
\usage{
sample_measurements(model, f, eta)
}
\arguments{
\item{model}{an object of type FrameModel.}

\item{f}{a data frame containing true parameter values.}

\item{eta}{numeric or vector, measurement noise overall or per isotopic measurement.}
}
\description{
Sample isotopic measurements from true parameter values
}
