% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeFRAME-package.R
\docType{package}
\name{TimeFRAME-package}
\alias{TimeFRAME-package}
\alias{TimeFRAME}
\title{The 'TimeFRAME' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2023). RStan: the R interface to Stan. R package version 2.21.8. https://mc-stan.org
}
